refactor: Separate stores views and URLs for consumer and owner backends
