+++ consumer-api/urls.py
@@ -3,7 +3,8 @@ from django.urls import include, path
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
 
 urlpatterns = [
-    path("api/v1/stores/", include("stores.urls")),      # 소비자용 공개 스토어 URL
+    # 소비자용 stores 앱 URL
+    path("api/v1/stores/", include("stores.urls")),
 
     # API 문서 관련 URL
     path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
