+++ stores/owner_urls.py
@@ -1,15 +1,11 @@
-# stores/urls.py
+# stores/owner_urls.py
 
 from django.urls import path
 
-from .views import MenuListView, OwnerStoreCreateView, StoreListView
+from .owner_views import OwnerStoreCreateView
 
 urlpatterns = [
-    # GET /api/v1/stores
-    path("", StoreListView.as_view(), name="store-list"),
-
-    # GET /api/v1/stores/{store_id}/menus
-    path("<int:store_id>/menus", MenuListView.as_view(), name="menu-list"),
-
+    # 사장님 전용 기능
+    # 가게 생성 API
     path("owner/stores/", OwnerStoreCreateView.as_view(), name="owner-store-create"),
 ]
