+++ stores/urls.py
@@ -1,15 +1,12 @@
 # stores/urls.py
 
 from django.urls import path
-
-from .views import MenuListView, OwnerStoreCreateView, StoreListView
+from .consumer_views import MenuListView, StoreListView
 
 urlpatterns = [
-    # GET /api/v1/stores
+    # 상점 목록
     path("", StoreListView.as_view(), name="store-list"),
 
-    # GET /api/v1/stores/{store_id}/menus
+    # 메뉴 조회
     path("<int:store_id>/menus", MenuListView.as_view(), name="menu-list"),
-
-    path("owner/stores/", OwnerStoreCreateView.as_view(), name="owner-store-create"),
 ]
